requirements.txt

flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
werkzeug
google-auth
requests
